chore: create code separation
