+++ Assets/furious-poultry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e74dc61178de14488aa34c65ead0715
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
