+++ Assets/furious-poultry/domain/ClampableIndex.cs.meta
