+++ Assets/furious-poultry/domain/furious-poultry.domain.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "furious-poultry.domain"
+}
