+++ Assets/furious-poultry/domain/furious-poultry.domain.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: af9f40918dfcb2846b5cb8e1992888e7
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
