+++ Assets/furious-poultry/unity/furious-poultry.unity.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "furious-poultry.unity",
+    "rootNamespace": "com.github.UnityWorkshop",
+    "references": [
+        "GUID:af9f40918dfcb2846b5cb8e1992888e7"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
