+++ Assets/furious-poultry/unity/furious-poultry.unity.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 64d3bc762e02cb64fa0faf28c4b5152e
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
